Label Utility 및 MovieData 데이터에 맞게 변환
