--- MovieRating/HomeTableViewController 2.swift
@@ -1,167 +1,0 @@
-//
-//  HomeTableViewController.swift
-//  MovieRating
-//
-//  Created by 혜리 on 2023/04/03.
-//
-
-import UIKit
-import SDWebImage
-
-class HomeTableViewController: UITableViewController {
-
-    private var movieTableView: UITableView = {
-        let view = UITableView()
-        view.backgroundColor = .white
-        view.rowHeight = 120
-        view.register(MovieCell.self, forCellReuseIdentifier: "MovieCell")
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-    
-    var movies = [Movie]()
-    var query: String?
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setup()
-        addViews()
-        setTableView()
-        setNavigationBar()
-        setConstraints()
-    }
-    
-    private func setup() {
-        view.backgroundColor = .white
-    }
-    
-    private func addViews() {
-        view.addSubview(movieTableView)
-    }
-    
-    private func setTableView() {
-        movieTableView.delegate = self
-        movieTableView.dataSource = self
-    }
-    
-    private func setNavigationBar() {
-        navigationController?.navigationBar.barTintColor = .white
-        navigationController?.navigationBar.tintColor = .black
-        navigationController?.navigationBar.titleTextAttributes = [.foregroundColor: UIColor.black]
-        
-        navigationItem.title = "홈"
-        navigationItem.searchController = UISearchController(searchResultsController: nil)
-        navigationItem.searchController?.searchBar.placeholder = "검색어를 입력하세요"
-        navigationItem.searchController?.searchBar.delegate = self
-    }
-    
-    private func setConstraints() {
-        NSLayoutConstraint.activate([
-            movieTableView.topAnchor.constraint(equalTo: view.topAnchor, constant: 10),
-            movieTableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
-            movieTableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
-            movieTableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor)
-        ])
-    }
-    
-    private func searchMovies(query: String) {
-
-        var urlComponents = URLComponents()
-        urlComponents.scheme = "https"
-        urlComponents.host = "api.themoviedb.org"
-        urlComponents.path = "/3/search/movie"
-        urlComponents.queryItems = [
-            URLQueryItem(name: "api_key", value: "73861d304f91be437d4465b52141a39b"),
-            URLQueryItem(name: "query", value: query),
-            URLQueryItem(name: "language", value: "ko-KR")
-        ]
-
-        guard let url = urlComponents.url else {
-              print("유효하지 않는 URL")
-              return
-          }
-
-        var request = URLRequest(url: url)
-            request.httpMethod = "GET"
-
-        let task = URLSession.shared.dataTask(with: url) { data, response, error in
-            guard let data = data, error == nil else {
-                print(error?.localizedDescription ?? "Unknown error")
-                return
-            }
-            do {
-                let decoder = JSONDecoder()
-                let response = try decoder.decode(Response.self, from: data)
-                self.movies = response.results
-                DispatchQueue.main.async {
-                    self.movieTableView.reloadData()
-                }
-            } catch {
-                print(error.localizedDescription)
-            }
-        }
-        task.resume()
-    }
-}
-
-
-extension HomeTableViewController {
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return movies.count
-    }
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "MovieCell", for: indexPath) as! MovieCell
-        let movie = movies[indexPath.row]
-        cell.titleAndYearLabel.text = "\(movie.title) (\(movie.releaseDate))"
-        if let posterPath = movie.posterPath {
-            let posterURL = URL(string: "https://image.tmdb.org/t/p/w500\(posterPath)")
-            cell.thumbnailImage.sd_setImage(with: posterURL, placeholderImage: UIImage(named: "placeholder"))
-        }
-
-        return cell
-    }
-}
-
-
-extension HomeTableViewController: UISearchBarDelegate {
-    
-    func searchBarTextDidEndEditing(_ searchBar: UISearchBar) {
-        self.movieTableView.reloadData()
-    }
-    
-    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-            
-        guard let searchTerm = searchBar.text, searchTerm.isEmpty == false else { return }
-        
-        DispatchQueue.main.async {
-            self.movieTableView.reloadData()
-        }
-        
-        if let text = searchBar.text {
-            query = text
-
-            searchMovies(query: query!)
-            print("\(text)")
-        }
-    }
-}
-
-
-struct Response: Codable {
-    let results: [Movie]
-}
-
-struct Movie: Codable {
-    let title: String
-    let releaseDate: String
-    let posterPath: String?
-    
-    enum CodingKeys: String, CodingKey {
-        case title = "title"
-        case releaseDate = "release_date"
-        case posterPath = "poster_path"
-    }
-}
